requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
alembic==1.13.1
aiosqlite==0.20.0
pydantic[email]==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
freezegun==1.5.1
loguru==0.7.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
asgi-lifespan==2.1.0
